howtouse_streamlit/requirements.txt
numpy
pillow
plotly
pandas
streamlit